% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KStestGIL.R
\name{KStestGIL}
\alias{KStestGIL}
\title{Kolmogorov-Smirnov Goodness of Fit Test for Distribution Generalized Inverse-Lindley}
\usage{
KStestGIL(x, gamaawal, alphacv = 0.05, eps = 1e-06)
}
\arguments{
\item{x}{is numeric variable}

\item{gamaawal}{is numeric variable and initial values}

\item{alphacv}{is level of significance and use 0,05}

\item{epsilon}{is error and use it epsilon = 0.000001}
}
\description{
Kolmogorov-Smirnov Goodness of Fit Test for Distribution Generalized Inverse-Lindley
}
\examples{
gamaawal = c(alpha,lambda)
KStestGIL(x, gamaawal, alphacv = 0.05, eps=0.000001)
}
